scrapers/tools/price_tracker/requirements.txt
gspread>=6.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
playwright>=1.40
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
